Move region catalog and encoder to case study
